hw3_rl_kalshi_vishal/requirements.txt
gymnasium==0.29.1
stable-baselines3==2.3.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
streamlit==1.36.0
requests==2.32.3
pydantic==2.8.2
python-dotenv==1.0.1
loguru==0.7.2
cryptography==43.0.1


